size_t to u32

+++ candle-nonzero/src/nonzero.cu
@@ -11,26 +11,26 @@ template <typename T> struct NonZeroOp {
 
 // count the number of non-zero elements in an array, to better allocate memory
 template <typename T>
-void count_nonzero(const T *d_in, const size_t N, size_t *h_out) {
+void count_nonzero(const T *d_in, const uint32_t N, uint32_t *h_out) {
   cub::TransformInputIterator<bool, NonZeroOp<T>, const T *> itr(
       d_in, NonZeroOp<T>());
   size_t temp_storage_bytes = 0;
   size_t *d_num_nonzero;
-  cudaMalloc((void **)&d_num_nonzero, sizeof(size_t));
+  cudaMalloc((void **)&d_num_nonzero, sizeof(uint32_t));
   cub::DeviceReduce::Sum(nullptr, temp_storage_bytes, itr, d_num_nonzero, N);
   void **d_temp_storage;
   cudaMalloc(&d_temp_storage, temp_storage_bytes);
   cub::DeviceReduce::Sum(d_temp_storage, temp_storage_bytes, itr,
                          d_num_nonzero, N);
-  cudaMemcpy(h_out, d_num_nonzero, sizeof(size_t), cudaMemcpyDeviceToHost);
+  cudaMemcpy(h_out, d_num_nonzero, sizeof(uint32_t), cudaMemcpyDeviceToHost);
   cudaFree(d_num_nonzero);
   cudaFree(d_temp_storage);
 }
 
 #define COUNT_NONZERO_OP(TYPENAME, RUST_NAME)                                  \
-  extern "C" size_t count_nonzero_##RUST_NAME(const TYPENAME *d_in,          \
-                                                size_t N) {                  \
-    size_t result;                                                           \
+  extern "C" uint32_t count_nonzero_##RUST_NAME(const TYPENAME *d_in,          \
+                                                uint32_t N) {                  \
+    uint32_t result;                                                           \
     count_nonzero(d_in, N, &result);                                           \
     return result;                                                             \
   }
@@ -49,10 +49,10 @@ COUNT_NONZERO_OP(uint8_t, u8)
 COUNT_NONZERO_OP(uint32_t, u32)
 COUNT_NONZERO_OP(int64_t, i64)
 
-__global__ void transform_indices(const size_t *temp_indices,
-                                  const size_t num_nonzero,
-                                  const size_t *dims, const size_t num_dims,
-                                  size_t *d_out) {
+__global__ void transform_indices(const uint32_t *temp_indices,
+                                  const uint32_t num_nonzero,
+                                  const uint32_t *dims, const uint32_t num_dims,
+                                  uint32_t *d_out) {
   const int idx = blockIdx.x * blockDim.x + threadIdx.x;
   if (idx < num_nonzero) {
     int temp_index = temp_indices[idx];
@@ -65,15 +65,15 @@ __global__ void transform_indices(const size_t *temp_indices,
 
 // get the indices of non-zero elements in an array
 template <typename T>
-void nonzero(const T *d_in, const size_t N, const size_t num_nonzero,
-             const size_t *dims, const size_t num_dims, size_t *d_out) {
+void nonzero(const T *d_in, const uint32_t N, const uint32_t num_nonzero,
+             const uint32_t *dims, const uint32_t num_dims, uint32_t *d_out) {
   cub::TransformInputIterator<bool, NonZeroOp<T>, const T *> itr(
       d_in, NonZeroOp<T>());
-  cub::CountingInputIterator<size_t> counting_itr(0);
-  size_t *out_temp;
-  size_t *num_selected_out;
-  cudaMalloc((void **)&out_temp, num_nonzero * sizeof(size_t));
-  cudaMalloc((void **)&num_selected_out, sizeof(size_t));
+  cub::CountingInputIterator<uint32_t> counting_itr(0);
+  uint32_t *out_temp;
+  uint32_t *num_selected_out;
+  cudaMalloc((void **)&out_temp, num_nonzero * sizeof(uint32_t));
+  cudaMalloc((void **)&num_selected_out, sizeof(uint32_t));
   size_t temp_storage_bytes = 0;
   cub::DeviceSelect::Flagged(nullptr, temp_storage_bytes, counting_itr, itr,
                              out_temp, num_selected_out, N);
@@ -93,8 +93,8 @@ void nonzero(const T *d_in, const size_t N, const size_t num_nonzero,
 
 #define NONZERO_OP(TYPENAME, RUST_NAME)                                        \
   extern "C" void nonzero_##RUST_NAME(                                         \
-      const TYPENAME *d_in, size_t N, size_t num_nonzero,                 \
-      const size_t *dims, size_t num_dims, size_t *d_out) {              \
+      const TYPENAME *d_in, uint32_t N, uint32_t num_nonzero,                 \
+      const uint32_t *dims, uint32_t num_dims, uint32_t *d_out) {              \
     nonzero(d_in, N, num_nonzero, dims, num_dims, d_out);                     \
   }
